python-dog-microservice/requirements.txt
grpcio
grpcio-tools
homi
